Refactor: Tightened contract on getVisibility.
Confirmed no code ever invokes `getVisibility` passing in `undefined`
for the `now` param. Modified the contract to state that you cannot
pass in `undefined` (the contract didn't make it clear what the
behaviour should have been if `undefined` was passed in anyway). Then
removed the code that tries to handle `undefined` being passed in.