Upgrade sanitize-html to 2.7.0
The previous version had a vulnerability regarding IDN domain names.
Based on a quick reading of the vulnerability, I don't think Nailbox was
affected (it relies on setting some attributes on iframes, but I believe
that Nailbox filters out the iframe tag), but it was an easy upgrade,
and it gets rid of the alerts on the vulnerability.

In commit 81440a29895f52fffbc6418555f9439dc19fb4ed I restricted
sanitize-html to be earlier than 1.19.3. The commit message says that
this was because the library used constructs that are not compatible
with the version of nodejs used at the time. It looks like the version
of nodejs used at the time is 5. The project now uses nodejs 12.

Testing done:
* Loaded an HTML e-mail and confirmed it renders properly on the UI.